Add files via upload
The following project is for a learning purpose and the data/resource is from University of Massachusetts and EliteDataScience.

Wine Taster project is to develop a ML program that can distinguish from one wine to another. The project is to introduce ourselves into the sklearn library, which contains many useful statistical tools and algorithms for machine learning, or to be specific in this case, prediction.

Since this project is a learning experience for ML with sklearn, we are weeding out a lot of exploraratory data analysis. I am planning on doing another project soon with sklearn and that will include extensive explorartory data analysis. Follow that project for exploraroty data analysis.